Split GenAI runtime components by role
