+++ aws-opentelemetry-distro/src/amazon/opentelemetry/distro/gen_ai/propagator/__init__.py
@@ -1,0 +1,6 @@
+# Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+# SPDX-License-Identifier: Apache-2.0
+
+from amazon.opentelemetry.distro.gen_ai.propagator.llm_context_propagator import GenAiLlmContextPropagator
+
+__all__ = ["GenAiLlmContextPropagator"]
