+++ aws-opentelemetry-distro/src/amazon/opentelemetry/distro/gen_ai/sampler/__init__.py
@@ -1,0 +1,6 @@
+# Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+# SPDX-License-Identifier: Apache-2.0
+
+from amazon.opentelemetry.distro.gen_ai.sampler.http_drop_sampler import GenAiHttpDropSampler
+
+__all__ = ["GenAiHttpDropSampler"]
