+++ aws-opentelemetry-distro/src/amazon/opentelemetry/distro/gen_ai/span_processor/__init__.py
@@ -1,0 +1,8 @@
+# Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+# SPDX-License-Identifier: Apache-2.0
+
+from amazon.opentelemetry.distro.gen_ai.span_processor.nested_client_span_processor import (
+    GenAiNestedClientSpanProcessor,
+)
+
+__all__ = ["GenAiNestedClientSpanProcessor"]
