+++ aws-opentelemetry-distro/src/amazon/opentelemetry/distro/instrumentation/langchain/callback_handler.py
@@ -11,7 +11,7 @@ from uuid import UUID
 from langchain_core.callbacks import BaseCallbackHandler
 from langchain_core.messages import convert_to_messages
 
-from amazon.opentelemetry.distro.gen_ai.http_span_collapsing import attach_llm_span_context
+from amazon.opentelemetry.distro.gen_ai._context import attach_llm_span_context
 from amazon.opentelemetry.distro.instrumentation.common.instrumentation_utils import (
     PROVIDER_MAP,
     DictWithLock,
