+++ aws-opentelemetry-distro/src/amazon/opentelemetry/distro/instrumentation/openai_agents/_processor.py
@@ -23,7 +23,7 @@ from agents.tracing.span_data import (
 from pydantic import BaseModel
 from typing_extensions import override
 
-from amazon.opentelemetry.distro.gen_ai.http_span_collapsing import attach_llm_span_context
+from amazon.opentelemetry.distro.gen_ai._context import attach_llm_span_context
 from amazon.opentelemetry.distro.instrumentation.common.instrumentation_utils import (
     PROVIDER_MAP,
     DictWithLock,
